+++ src/main/java/core/Core.java
@@ -1,5 +1,7 @@
 package core;
 
+import core.gameObjects.*;
+
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Observable;
