+++ src/main/java/core/gameObjects/GameObject.java
@@ -1,4 +1,6 @@
-package core;
+package core.gameObjects;
+
+import core.Position;
 
 import java.util.Observable;
 import java.util.Observer;
