+++ src/main/java/core/gameObjects/IGameObject.java
@@ -1,4 +1,6 @@
-package core;
+package core.gameObjects;
+
+import core.Position;
 
 public interface IGameObject {
