+++ src/main/java/core/gameObjects/Missile.java
@@ -1,4 +1,6 @@
-package core;
+package core.gameObjects;
+
+import core.Position;
 
 public class Missile extends FlightObject {
