+++ src/main/java/core/gameObjects/UFO.java
@@ -1,4 +1,4 @@
-package core;
+package core.gameObjects;
 
 public class UFO extends FlightObject {
